fix: catch error in AuthConfirmCallProvider, swap AuthConfirmCallProvider and AuthConfirmSmsProvider names
